(build) osx/bsd/linux cleanup
Interestingly enough, it seems posix_spawnpt has rather different
behavior across the line when it comes to nonblocking/cloexec.
FBSD fails if any unspecified flags are used, while-as OSX only
accepts the nonblock etc. at creation time - and linux seems happy
with whatever.

FossilOrigin-Name: 4980cb78883add16079b5646e59d15450f93cdc5ec0bccb591587df5cfcd6a40

+++ src/frameserver/terminal/default/tsm/tsm_vte.c
@@ -434,7 +434,8 @@ SHL_EXPORT
 void tsm_vte_set_color(struct tsm_vte *vte,
 	enum vte_color ind, const uint8_t rgb[3])
 {
-	if (ind >= VTE_COLOR_NUM || ind < 0)
+
+	if (ind >= VTE_COLOR_NUM || (int)ind < 0)
 		return;
 
 	vte->palette[ind][0] = rgb[0];
